server/requirements.txt
fastapi
uvicorn[standard]
pydantic>=2.0
pydantic-settings
python-dotenv
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
email-validator